pyproject.toml
[tool.poetry]
name = "edpb"
version = "0.1.0"
description = "Elite Dangerous Price Bot, for Discord"
authors = ["Bruno Marques <bruno@bmarques.net>"]

[tool.poetry.dependencies]
python = "^3.9"
"discord.py" = "^2.0.0"
python-dotenv = "^0.20.0"
peewee = "^3.15.1"
apsw = "^3.39.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
